MinGW compatibility: 64-bit unsigned integer constants must be suffixed with ULL on GCC
git-svn-id: http://miranda.googlecode.com/svn/trunk@7910 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/src/modules/database/dbtime.c
@@ -185,7 +185,7 @@ static int TimestampToStringW(WPARAM wParam,LPARAM lParam)
 	int destCharsLeft, dateTimeStrLen;
 
 	//this huge number is the difference between 1970 and 1601 in seconds
-	liFiletime.QuadPart = (11644473600i64+(__int64)(DWORD)TimestampToLocal(wParam,0))*10000000;
+	liFiletime.QuadPart=((__int64)11644473600+(__int64)(DWORD)TimestampToLocal(wParam,0))*10000000;
 	filetime.dwHighDateTime = liFiletime.HighPart;
 	filetime.dwLowDateTime = liFiletime.LowPart;
 	FileTimeToSystemTime(&filetime,&st);
